feat: wrap sidebar tabs and menu in single
